📦 NEW: Get provider function in core
